requirements.txt
django==3.0.4
djangorestframework==3.11.0
pytest==5.3.5
pytest-django==3.8.0
pytest-cov==2.8.1
